requirements.txt
fastapi>=0.100.0
uvicorn[standard]
python-dotenv>=1.0.0
authlib>=1.2.0
python-jose[cryptography]>=3.3.0
sqlmodel>=0.0.8
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
alembic>=1.11.0
python-multipart>=0.0.6
asyncpg